FUSE/requirements.txt
anthropic
python-dotenv
fastapi==0.115.13
uvicorn==0.34.3
